+++ CHANGELOG.md
@@ -23,6 +23,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - A failed startup set global commands call will no longer lead to it retrying on the next startup.
+- Component level check errors not being caught when executing interactions.
+- The internal state for whether a SlashContext has had it's initial response made or not not being set by
+  SlashContext.mark_not_found.
+- Client.on_interaction_create_request not awaiteing the client level checks.
 
 ## [2.0.0a3.post1] - 2021-09-10
 ### Changed